+++ selfdrive/ui/onroad/augmented_road_view.py
@@ -3,12 +3,12 @@ import pyray as rl
 
 from cereal import log
 from msgq.visionipc import VisionStreamType
-from openpilot.system.ui.lib.ui_state import ui_state, UIStatus, UI_BORDER_SIZE
-from openpilot.system.ui.onroad.alert_renderer import AlertRenderer
-from openpilot.system.ui.onroad.driver_state import DriverStateRenderer
-from openpilot.system.ui.onroad.hud_renderer import HudRenderer
-from openpilot.system.ui.onroad.model_renderer import ModelRenderer
-from openpilot.system.ui.widgets.cameraview import CameraView
+from openpilot.selfdrive.ui.ui_state import ui_state, UIStatus, UI_BORDER_SIZE
+from openpilot.selfdrive.ui.onroad.alert_renderer import AlertRenderer
+from openpilot.selfdrive.ui.onroad.driver_state import DriverStateRenderer
+from openpilot.selfdrive.ui.onroad.hud_renderer import HudRenderer
+from openpilot.selfdrive.ui.onroad.model_renderer import ModelRenderer
+from openpilot.selfdrive.ui.onroad.cameraview import CameraView
 from openpilot.system.ui.lib.application import gui_app
 from openpilot.common.transformations.camera import DEVICE_CAMERAS, DeviceCameraConfig, view_frame_from_device_frame
 from openpilot.common.transformations.orientation import rot_from_euler
@@ -175,6 +175,7 @@ if __name__ == "__main__":
   print("***press space to switch camera view***")
   try:
     for _ in gui_app.render():
+      ui_state.update()
       if rl.is_key_released(rl.KeyboardKey.KEY_SPACE):
         is_wide = road_camera_view.stream_type == VisionStreamType.VISION_STREAM_WIDE_ROAD
         road_camera_view.switch_stream(VisionStreamType.VISION_STREAM_ROAD if is_wide else VisionStreamType.VISION_STREAM_WIDE_ROAD)
