+++ selfdrive/ui/onroad/driver_state.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pyray as rl
 from dataclasses import dataclass
+from openpilot.selfdrive.ui.ui_state import ui_state, UI_BORDER_SIZE
 from openpilot.system.ui.lib.application import gui_app
-from openpilot.system.ui.lib.ui_state import ui_state, UI_BORDER_SIZE
 
 
 # Default 3D coordinates for face keypoints as a NumPy array
