+++ selfdrive/ui/onroad/hud_renderer.py
@@ -1,7 +1,7 @@
 import pyray as rl
 from dataclasses import dataclass
 from cereal.messaging import SubMaster
-from openpilot.system.ui.lib.ui_state import ui_state, UIStatus
+from openpilot.selfdrive.ui.ui_state import ui_state, UIStatus
 from openpilot.system.ui.lib.application import gui_app, FontWeight
 from openpilot.common.conversions import Conversions as CV
