+++ selfdrive/ui/onroad/model_renderer.py
@@ -4,8 +4,8 @@ import pyray as rl
 from cereal import messaging, car
 from dataclasses import dataclass, field
 from openpilot.common.params import Params
+from openpilot.selfdrive.ui.ui_state import ui_state
 from openpilot.system.ui.lib.application import DEFAULT_FPS
-from openpilot.system.ui.lib.ui_state import ui_state
 from openpilot.system.ui.lib.shader_polygon import draw_polygon
